Fix c89 bugs reported by Cory Stup.
Others may remain.  I wasn't able to get gcc --std=c89 to build libevent
at all, so I don't know what compiler the original reporter is using here.

Note that this change requires us to disable the part of our rpc code
that uses variadic macros when using a non-gcc compiler.  This is a
problem if we want our rpc api to be portable.

git-svn-id: https://levent.svn.sourceforge.net/svnroot/levent/trunk/libevent@1231 ce677c24-8c1a-0410-9497-a30ef3a96221