Activate fd events in a pseudorandom order on older backends.
New backends like poll and kqueue and so on add fds to the queue in
the order that they are triggered.  But the select backend currently
activates low-numbered fds first, whereas the poll and win32 backends
currently favor whatever fds have been on for the longest.  This is no
good for fairness.

git-svn-id: https://levent.svn.sourceforge.net/svnroot/levent/branches/patches-1.4@1327 ce677c24-8c1a-0410-9497-a30ef3a96221