Make epoll use less RAM.
We do this by not allocating the maximum epoll_event array for the epoll
backend at startup.  Instead, we start out accepting 32 events at a time, and
double the array's size when it seems that the OS is generating events faster
than we're requesting them.  This saves up to 374K per epoll-based
event_base.  Resolves bug 2839240.

git-svn-id: https://levent.svn.sourceforge.net/svnroot/levent/trunk@1428 ce677c24-8c1a-0410-9497-a30ef3a96221