Make the bufferevent_connect_fail test faster on OSX.
It seems that connecting to a listener that is bound but not accepting
or listening doesn't give a 'connection refused' error on OSX, but
rather makes the connect() time out after 75 seconds.  I couldn't find
any way to make the timout shorter.  Fortunately, closing the listener
after a second or so makes the desired error occur after another
second or so.

git-svn-id: https://levent.svn.sourceforge.net/svnroot/levent/trunk@1457 ce677c24-8c1a-0410-9497-a30ef3a96221