Fix a miscalculated realloc() size in win32select.c.
This bug was introduced by the code to make the backend able to safely release the base lock while calling select().

Also, we change win32select.c to the same 32-fds-to-start default as the rest of the backends, so that the main/many_events test can test it.  It was at 64-to-start, so the test wasn't hitting it.

git-svn-id: https://levent.svn.sourceforge.net/svnroot/levent/trunk@1513 ce677c24-8c1a-0410-9497-a30ef3a96221